pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-detection-platform"
version = "0.1.0"
description = "Sigma-based detection-as-code platform with Claude-powered alert triage and response playbooks."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Venkata Prasanth Divyakolu" },
]
keywords = ["security", "detection-engineering", "sigma", "siem", "anthropic", "claude", "soar"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Security",
]

dependencies = [
    "duckdb>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "anthropic>=0.39.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "types-PyYAML",
]

[project.scripts]
cdp = "cdp.cli:app"

[project.urls]
Repository = "https://github.com/dprasanthv/claude-detection-platform"

[tool.setuptools.packages.find]
where = ["."]
include = ["cdp*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "integration: tests that hit external services (skipped by default)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Typer / Click are designed to be called as argument defaults; B008 is a
# false positive for this idiom and the docs explicitly recommend it.
"cdp/cli.py" = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
